# docs/openapi/cart-service-api.yml
openapi: 3.0.3   #openapi spec version
info:  #api hakkında genel bilgiler saglar
  title: Cart Service Api
  description: |
    This API manage cart service operations.
  version: 1.0.0

servers:
  - url: http://localhost:8087
    description: Local development server

#apinin tüm endpointlerini açıklar
paths:
  /api/v1/carts/{customerId}:      #paths object
    post:             #path item object
      tags:
        - carts
      summary: Add Product to Cart
      description: Add an item to the cart
      operationId: addToCart
      parameters:
        - $ref: '#/components/parameters/CustomerIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CartItemRequest'}
      responses:
        '201':
          description: Product added successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CartItemResponse'}

    get:
      tags:
        - carts
      summary: Get the customer's cart
      description: Retrieves the current customer's cart state
      operationId: getCartByCustomerId
      parameters:     #parametre nesnelerini tanımla.
        - $ref: '#/components/parameters/CustomerIdPathParam'
      responses:
        '200':
          description: Cart details retrieved successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CartResponse'}
        '404':
          description: Cart not found

  /api/v1/carts/items/{productId}:
    delete:
      tags:
        - carts
      summary: Delete product from Cart
      operationId: deleteProduct
      parameters:
        - $ref: '#/components/parameters/ProductIdPathParam'
      responses:
        '200':
          description: Product released from cart successfully

  /api/v1/carts/update-quantity/{productId}:
    patch:
      tags:
        - carts
      summary: Update product quantity
      description: Updates the quantity of the selected product.
      operationId: updateProductQuantity
      parameters:
        - $ref: '#/components/parameters/ProductIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateQuantityRequest'}
      responses:
        '200':
          description: Product quantity updated successfully
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CartItemResponse'}


components:
  # Reusable schemas (data models)
  schemas:
    CartItemRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId: {type: string, format: uuid}
        quantity: {type: integer, minimum: 1 }
    CartItemResponse:
      type: object
      properties:
        productId: {type: string, format: uuid}
        quantity: { type: integer }
        unitPrice: {type: number, format: decimal}

    CartItem:
      type: object
      properties:
        productId: {type: string, format: uuid}
        quantity: {type: integer}
        unitPrice: {type: number, format: decimal}
    CartResponse:
      type: object
      properties:
        cartTotalPrice: {type: number, format: decimal}
        items: {type: array, minItems: 0, items: {$ref: '#/components/schemas/CartItem'} }

    UpdateQuantityRequest:
      type: object
      properties:
        quantity: { type: integer, minimum: 1 }

 # Reusable path, query, header and cookie parameters
  parameters:
    CustomerIdPathParam:
      name: customerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: "d7cfe047-e53b-49a0-a486-d9c6ae27e8f8"

    ProductIdPathParam:
      name: productId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: "d9aefa38-923e-432a-b0a0-5d0bd8822a25"